requirements.txt
aio-pika==8.2.5
fastapi==0.87.0
uvicorn==0.20.0
pydantic~=1.10.2
pyyaml==6.0
python-dotenv==0.21.0
email-validator==1.3.0